requirements.txt
# CORE KERNEL (The Brain)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
google-genai>=0.3.0

# MEMORY & DATABASE (The Librarian)
sqlalchemy>=2.0.25
chromadb>=0.4.22
# For generic SQL support (SQLite is built-in, but this helps)
aiosqlite>=0.19.0

# AGENT TOOLS (The Hands)
playwright>=1.41.0
aiohttp>=3.9.1
requests>=2.31.0
beautifulsoup4>=4.12.3

# FRONTEND (The Face)
streamlit>=1.31.0
watchdog>=4.0.0

# UTILITIES
pyyaml>=6.0.1
tenacity>=8.2.3  # For retrying failed agents